ref(discover): Move table detection function to Discover dataset
Move the table detection code from the Discover entity to the dataset,
where it is now actually being used. This is a straight copy and paste -
cleaning up this method will be done as a follow up.
